Configure Prettier
This commit adds Prettier to the project for formatting. A "format"
script is also provided to format from the command line. Most editors
can format on save, but I still find this to be useful.

The default formatting options are not changed. I don't agree with
Prettier on everything. I wish it used single quotes, for example.
Still, consistency with other projects is important and beginning down
the road of discussing different configuration options just seems like a
waste of time.

I haven't had much luck with Svelte plugins, so this commit doesn't add
support for formatting Svelte files. We can revisit Svelte formatting in
the future.